requirements.txt
django~=3.1.5
djangorestframework~=3.12.2
pylint~=2.6.0
django-environ~=0.4.5
psycopg2~=2.8.6
autopep8~=1.5.4
pylint-django~=2.4.2
requests~=2.25.1